+++ docs/changelog.md
@@ -89,6 +89,86 @@
 
 ## 更新说明详情
 
+1.18.38 fix: fix stock_hk_index_daily_sina interface
+
+	1. 修复 stock_hk_index_daily_sina 接口
+
+1.18.37 fix: fix fund_etf_dividend_sina interface
+
+	1. 修复 fund_etf_dividend_sina 接口
+
+1.18.36 fix: fix stock_zygc_em interface
+
+	1. 修复 stock_zygc_em 接口
+
+1.18.35 fix: fix stock_ipo_ths interface
+
+	1. 修复 stock_ipo_ths 接口
+
+1.18.34 fix: fix fund_fh_em interface
+
+	1. 修复 fund_fh_em 接口
+
+1.18.33 fix: fix futures_spot_price_daily interface
+
+	1. 修复 futures_spot_price_daily 接口
+
+1.18.32 fix: fix fund_new_found_ths interface
+
+	1. 修复 fund_new_found_ths 接口
+
+1.18.31 fix: fix stock_hk_dividend_payout_em interface
+
+	1. 修复 stock_hk_dividend_payout_em 接口
+
+1.18.30 fix: fix futures_inventory_99 interface
+
+	1. 修复 futures_inventory_99 接口
+
+1.18.29 fix: fix stock_individual_basic_info_xq interface
+
+	1. 修复 stock_individual_basic_info_xq  接口
+
+1.18.28 fix: fix futures_spot_price_daily interface
+
+	1. 修复 futures_spot_price_daily 接口
+
+1.18.27 fix: fix futures_comm_js interface
+
+	1. 修复 futures_comm_js 接口
+
+1.18.26 fix: fix fund_fh_em interface
+
+	1. 修复 fund_fh_em 接口
+	2. 修复 stock_gdfx_free_holding_change_em 接口
+
+1.18.25 fix: fix qdii_e_comm_jsl interface
+
+	1. 修复 qdii_e_comm_jsl 接口
+	2. 修复 stock_rank_xstp_ths 接口
+
+1.18.24 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
+1.18.23 fix: fix fund_fee_em interface
+
+	1. 修复 fund_fee_em 接口
+	2. 修复 fund_etf_category_ths 接口
+
+1.18.22 fix: fix bond_gb_zh_sina interface
+
+	1. 修复 bond_gb_zh_sina 接口
+	2. 修复 bond_gb_us_sina 接口
+
+1.18.21 fix: fix get_receipt interface
+
+	1. 修复 get_receipt 接口
+
+1.18.20 fix: fix stock_individual_spot_xq interface
+
+	1. 修复 stock_individual_spot_xq 接口
+
 1.18.19 fix: fix GitHub action
 
 	1. 修复配置文件
@@ -610,7 +690,7 @@
 	1. 修复 stock_info_sz_delist 接口
 
 1.16.98 fix: fix fund_open_fund_info_em interface
-	
+
 	1. 修复 fund_open_fund_info_em 接口
 
 1.16.97 fix: fix fund_open_fund_info_em interface
@@ -618,7 +698,7 @@
 	1. 修复 fund_open_fund_info_em 接口
 
 1.16.96 fix: fix stock_individual_spot_xq interface
-	
+
 	1. 修复 stock_individual_spot_xq 接口
 	2. 修复 stock_cyq_em 接口
 
@@ -6105,6 +6185,44 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.18.38 fix: fix stock_hk_index_daily_sina interface
+
+1.18.37 fix: fix fund_etf_dividend_sina interface
+
+1.18.36 fix: fix stock_zygc_em interface
+
+1.18.35 fix: fix stock_ipo_ths interface
+
+1.18.34 fix: fix fund_fh_em interface
+
+1.18.33 fix: fix futures_spot_price_daily interface
+
+1.18.32 fix: fix fund_new_found_ths interface
+
+1.18.31 fix: fix stock_hk_dividend_payout_em interface
+
+1.18.30 fix: fix futures_inventory_99 interface
+
+1.18.29 fix: fix stock_individual_basic_info_xq interface
+
+1.18.28 fix: fix futures_spot_price_daily interface
+
+1.18.27 fix: fix futures_comm_js interface
+
+1.18.26 fix: fix fund_fh_em interface
+
+1.18.25 fix: fix qdii_e_comm_jsl interface
+
+1.18.24 fix: fix stock_individual_spot_xq interface
+
+1.18.23 fix: fix fund_fee_em interface
+
+1.18.22 fix: fix bond_gb_zh_sina interface
+
+1.18.21 fix: fix get_receipt interface
+
+1.18.20 fix: fix stock_individual_spot_xq interface
 
 1.18.19 fix: fix GitHub action